+++ app/TestService.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TestService;
+
+class TestService
+{
+
+    public function testMethod($param = "") {}
+}