+++ CHANGELOG.md
@@ -8,6 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.1] - 2019-04-20
+
+### Added
+
+### Fixed
+
+
+## [1.2.0] - 2019-04-19
+
+### Added
+
+- Added `resource` controller
+- Added `craft:resource`
+- Added `resource` tests
+
 ## [1.1.2] - 2019-04-18
 
 ### Fixed